+++ .gitignore
@@ -3,5 +3,7 @@
 WeatherApp.xcodeproj/project.xcworkspace/xcuserdata/patricia.xcuserdatad/UserInterfaceState.xcuserstate
 WeatherApp.xcodeproj/project.xcworkspace/xcuserdata/patricia.xcuserdatad/UserInterfaceState.xcuserstate
 WeatherApp.xcodeproj/project.xcworkspace/xcuserdata/patricia.xcuserdatad/UserInterfaceState.xcuserstate
+WeatherApp.xcodeproj/xcuserdata/patricia.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
 *.xcuserstate
 *.xcuserstate
+WeatherApp.xcodeproj/xcuserdata/patricia.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
