--- Charts/Charts/ChartsApp.swift
@@ -1,17 +1,0 @@
-//
-//  ChartsApp.swift
-//  Charts
-//
-//  Created by Patricia Costin on 23.01.2024.
-//
-
-import SwiftUI
-
-@main
-struct ChartsApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
