--- Test/Test/TestApp.swift
@@ -1,17 +1,0 @@
-//
-//  TestApp.swift
-//  Test
-//
-//  Created by Patricia Costin on 24.01.2024.
-//
-
-import SwiftUI
-
-@main
-struct TestApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
